made few changes

+++ src/main/java/com/ktk/ktkapp/repos/kiosk/kioskComponentRepo.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
 public interface kioskComponentRepo extends JpaRepository<kioskComponentModel, kioskComponentModel.KioskComponentId> {
-    Optional<kioskComponentModel>findByKioskIdAndComponentId(Long kioskId, Long componentId);
+    Optional<kioskComponentModel> findByKioskIdAndComponentId(Integer kioskId, Integer componentId);
 }
